+++ example/connectivity/js/main.js
@@ -2,8 +2,7 @@ require.config( {
 	paths: {
 		text: "../../../bower/requirejs-text/text",
 		backbone: '../../../bower/backbone/backbone',
-		lodash: '../../../bower/lodash/dist/lodash',
-		underscore: '../../../bower/lodash/dist/lodash.underscore',
+		lodash: '../../../bower/lodash/lodash',
 		mockjax: '../../../bower/jquery-mockjax/jquery.mockjax',
 		machina: '/lib/machina',
 		'machina.postal': '../../../bower/machina.postal/lib/machina.postal',
@@ -11,6 +10,11 @@ require.config( {
 		'postal.diags': '../../../bower/postal.diagnostics/lib/postal.diagnostics.min',
 		jquery: '../../../bower/jquery/jquery',
 		conduitjs: '../../../bower/conduitjs/lib/conduit'
+	},
+	map: {
+		"*": {
+			"underscore": "lodash"
+		}
 	},
 	shim: {
 		mockjax: [ 'jquery' ],