+++ gulpfile.js
@@ -1,59 +1,109 @@
 var gulp = require( "gulp" );
-var fileImports = require( "gulp-imports" );
-var header = require( "gulp-header" );
-var beautify = require( "gulp-beautify" );
-var hintNot = require( "gulp-hint-not" );
-var uglify = require( "gulp-uglify" );
-var rename = require( "gulp-rename" );
-var plato = require( "gulp-plato" );
 var gutil = require( "gulp-util" );
-var pkg = require( "./package.json" );
-var express = require( "express" );
+var rename = require( "gulp-rename" );
+var uglify = require( "gulp-uglify" );
+var _ = require( "lodash" );
+var eslint = require( "gulp-eslint" );
+var jscs = require( "gulp-jscs" );
+var gulpChanged = require( "gulp-changed" );
+var webpack = require( "gulp-webpack" );
 var path = require( "path" );
+var karma = require( "karma" );
+var mocha = require( "gulp-spawn-mocha" );
+var sourcemaps = require( "gulp-sourcemaps" );
+var express = require( "express" );
 var open = require( "open" );
 var port = 3080;
-var allSrcFiles = "./src/**/*.js";
 
-var banner = [ "/**",
-    " * <%= pkg.name %> - <%= pkg.description %>",
-    " * Author: <%= pkg.author %>",
-    " * Version: v<%= pkg.version %>",
-    " * Url: <%= pkg.homepage %>",
-    " * License(s): <% pkg.licenses.forEach(function( license, idx ){ %><%= license.type %><% if(idx !== pkg.licenses.length-1) { %>, <% } %><% }); %>",
-    " */",
-    "" ].join( "\n" );
+gulp.task( "default", [ "build" ] );
 
-gulp.task( "combine", function() {
-	gulp.src( [ "./src/machina.js" ] )
-		.pipe( header( banner, {
-			pkg: pkg
-		} ) )
-		.pipe( fileImports() )
-		.pipe( hintNot() )
-		.pipe( beautify( {
-			indentSize: 4
-		} ) )
-		.pipe( gulp.dest( "./lib/" ) )
-		.pipe( gulp.dest( "./lib/" ) )
+gulp.task( "build", [ "format" ], function() {
+	return gulp.src( "src/machina.js" )
+		.pipe( webpack( require( "./webpack.config.js" ) ) )
+		.pipe( gulp.dest( "lib/" ) )
+		.pipe( sourcemaps.init( { loadMaps: true } ) )
 		.pipe( uglify( {
+			preserveComments: "license",
 			compress: {
+				/*eslint-disable */
 				negate_iife: false
+				/*eslint-enable */
 			}
 		} ) )
-		.pipe( header( banner, {
-			pkg: pkg
-		} ) )
 		.pipe( rename( "machina.min.js" ) )
-		.pipe( gulp.dest( "./lib/" ) );
+		.pipe( sourcemaps.write( "./" ) )
+		.pipe( gulp.dest( "lib/" ) );
 } );
 
-gulp.task( "default", function() {
-	gulp.run( "combine" );
+function runTests( options, done ) {
+	var server = new karma.Server( _.extend( {
+		configFile: path.join( __dirname, "/karma.conf.js" ),
+		singleRun: true
+
+		// no-op keeps karma from process.exit'ing gulp
+	}, options ), done || function() {} );
+
+	server.start();
+}
+
+gulp.task( "test", [ "build" ], function( done ) {
+	runTests( { reporters: [ "spec" ] }, function( err ) {
+		if ( err !== 0 ) {
+			// Exit with the error code
+			process.exit( err );
+		} else {
+			done( null );
+		}
+	} );
 } );
 
-gulp.task( "report", function() {
-	gulp.src( "./lib/machina.js" )
-		.pipe( plato( "report" ) );
+gulp.task( "coverage", [ "build" ], function( done ) {
+	runTests( {}, function( err ) {
+		if ( err !== 0 ) {
+			// Exit with the error code
+			process.exit( err );
+		} else {
+			done( null );
+		}
+	} );
+} );
+
+gulp.task( "mocha", [ "build" ], function() {
+	return gulp.src( [ "spec/**/*.spec.js" ], { read: false } )
+		.pipe( mocha( {
+			require: [ "spec/helpers/node-setup.js" ],
+			reporter: "spec",
+			colors: true,
+			inlineDiffs: true,
+			debug: false
+		} ) )
+		.on( "error", console.warn.bind( console ) );
+} );
+
+gulp.task( "lint", function() {
+	return gulp.src( [ "src/**/*.js", "spec/**/*.spec.js" ] )
+	.pipe( eslint() )
+	.pipe( eslint.format() )
+	.pipe( eslint.failOnError() );
+} );
+
+gulp.task( "format", [ "lint" ], function() {
+	return gulp.src( [ "*.js", "{spec,src}/**/*.js" ] )
+		.pipe( jscs( {
+			configPath: ".jscsrc",
+			fix: true
+		} ) )
+		.on( "error", function( error ) {
+			gutil.log( gutil.colors.red( error.message ) );
+			this.end();
+		} )
+		.pipe( gulpChanged( ".", { hasChanged: gulpChanged.compareSha1Digest } ) )
+		.pipe( gulp.dest( "." ) );
+} );
+
+gulp.task( "watch", function() {
+	gulp.watch( "src/**/*", [ "default" ] );
+	gulp.watch( "{lib,spec}/**/*", [ "mocha" ] );
 } );
 
 var createServers = function( port ) {
@@ -71,71 +121,12 @@ var createServers = function( port ) {
 
 var servers;
 
-gulp.task( "server", [ "combine", "report" ], function() {
+gulp.task( "server", [ "build" ], function() {
 	if ( !servers ) {
 		servers = createServers( port );
 	}
 
 	open( "http://localhost:" + port + "/index.html" );
-} );
-
-var mocha = require( "gulp-spawn-mocha" );
-gulp.task( "mocha", function() {
-	return gulp.src( [ "spec/**/*.spec.js" ], { read: false } )
-		.pipe( mocha( {
-			require: [ "spec/helpers/node-setup.js" ],
-			reporter: "spec",
-			colors: true,
-			inlineDiffs: true,
-			debug: false
-		} ) )
-		.on( "error", console.warn.bind( console ) );
-} );
-
-gulp.task( "watch", [ "default", "mocha" ], function() {
-	gulp.watch( "src/**/*", [ "default" ] );
-	gulp.watch( "{lib,spec}/**/*", [ "mocha" ] );
-} );
-
-var jscs = require( "gulp-jscs" );
-var gulpChanged = require( "gulp-changed" );
-
-gulp.task( "format", [ "jshint" ], function() {
-	return gulp.src( [ "**/*.js", "!node_modules/**" ] )
-		.pipe( jscs( {
-			configPath: ".jscsrc",
-			fix: true
-		} ) )
-		.pipe( gulpChanged( ".", { hasChanged: gulpChanged.compareSha1Digest } ) )
-		.pipe( gulp.dest( "." ) );
-} );
-
-var jshint = require( "gulp-jshint" );
-var stylish = require( "jshint-stylish" );
-
-gulp.task( "jshint", function() {
-	return gulp.src( allSrcFiles )
-		.on( "error", function( error ) {
-			gutil.log( gutil.colors.red( error.message + " in " + error.fileName ) );
-			this.end();
-		} )
-		.pipe( jshint() )
-		.pipe( jshint.reporter( stylish ) )
-		.pipe( jshint.reporter( "fail" ) );
-} );
-
-gulp.task( "coverage", [ "format" ], function( cb ) {
-	gulp.src( [ allSrcFiles ] )
-		.pipe( istanbul() ) // Covering files
-		.pipe( istanbul.hookRequire() ) // Force `require` to return covered files
-		.on( "finish", function() {
-			gulp.src( [ "./spec/helpers/node-setup.js", allTestFiles ] )
-				.pipe( mocha() )
-				.pipe( istanbul.writeReports() ) // Creating the reports after tests runned
-				.on( "end", function() {
-					process.exit();
-				} );
-		} );
 } );
 
 gulp.task( "show-coverage", function() {