+++ spec/hierarchical.BehavioralFsm.spec.js
@@ -1,8 +1,12 @@
+var _ = require( "lodash" );
+var sinon = require( "sinon" );
+var machina = require( "../lib/machina.js" );
+var hierarchical = require( "./helpers/hierarchicalFsm.js" )( machina );
+
 describe( "Hierarchical machina.BehavioralFsm", function() {
 	describe( "when creating a hierarchy", function() {
 		var crosswalk;
 		var events = [];
-		var resetHandled = false;
 		var client = {};
 		before( function() {
 			crosswalk = hierarchical.behavioralCrosswalkFactory( {
@@ -18,6 +22,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 		} );
 		it( "should report correct starting state for parent FSM", function() {
 			client.__machina__.crosswalk.state.should.equal( "vehiclesEnabled" );
+			crosswalk.compositeState( client ).should.equal( "vehiclesEnabled.green" );
 		} );
 		it( "should emit a 'pedestrians - do not walk' event", function() {
 			events[ 3 ].should.eql( { name: "pedestrians", data: { status: "Do Not Walk", client: client } } );
@@ -26,7 +31,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 			client.__machina__[ "vehicle-signal" ].state.should.equal( "green" );
 		} );
 		it( "should issue reset input to child FSM of active parent state", function() {
-			events[ 5 ].should.eql( {
+			events[ 6 ].should.eql( {
 				name: "handling",
 				data: {
 					inputType: "_reset",
@@ -38,7 +43,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 			} );
 		} );
 		it( "should emit a 'vehicles - green' event", function() {
-			events[ 8 ].should.eql( { name: "vehicles", data: { status: "green", client: client } } );
+			events[ 9 ].should.eql( { name: "vehicles", data: { status: "green", client: client } } );
 		} );
 		it( "should not be listening to any events from child FSM of inactive parent state", function() {
 			_.any( events, function( item ) {
@@ -50,7 +55,6 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 		describe( "and the input originates from timer in child", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			var client = {};
 			before( function() {
 				this.clock = sinon.useFakeTimers();
@@ -91,6 +95,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 					}
 				} );
 				client.__machina__[ "vehicle-signal" ].state.should.equal( "green-interruptible" );
+				crosswalk.compositeState( client ).should.equal( "vehiclesEnabled.green-interruptible" );
 			} );
 			it( "should not change parent FSM's state", function() {
 				client.__machina__.crosswalk.state.should.equal( "vehiclesEnabled" );
@@ -99,7 +104,6 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 		describe( "and the input originates from parent FSM", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			var client = {};
 			before( function() {
 				this.clock = sinon.useFakeTimers();
@@ -136,6 +140,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 					}
 				} );
 				client.__machina__[ "vehicle-signal" ].state.should.equal( "yellow" );
+				crosswalk.compositeState( client ).should.equal( "vehiclesEnabled.yellow" );
 			} );
 			it( "should emit a 'vehicles - yellow' event", function() {
 				events[ 2 ].should.eql( { name: "vehicles", data: { status: "yellow", client: client } } );
@@ -144,7 +149,6 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 		describe( "and input isn't handled in child, but parent instead", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			var client = {};
 			before( function() {
 				this.clock = sinon.useFakeTimers();
@@ -187,12 +191,13 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 					}
 				} );
 				client.__machina__.crosswalk.state.should.equal( "pedestriansEnabled" );
+				crosswalk.compositeState( client ).should.equal( "pedestriansEnabled.walking" );
 			} );
 			it( "should emit a 'vehicles - red' event", function() {
 				events[ 2 ].should.eql( { name: "vehicles", data: { status: "red", client: client } } );
 			} );
 			it( "should have child FSM of active parent state handle _reset input", function() {
-				events[ 4 ].should.eql( {
+				events[ 5 ].should.eql( {
 					name: "handling",
 					data: {
 						inputType: "_reset",
@@ -204,13 +209,12 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 				} );
 			} );
 			it( "should emit a 'pedestrians - walk' event", function() {
-				events[ 7 ].should.eql( { name: "pedestrians", data: { status: "Walk", client: client } } );
+				events[ 8 ].should.eql( { name: "pedestrians", data: { status: "Walk", client: client } } );
 			} );
 		} );
 		describe( "and parent FSM transitions into previously held state", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			var client = {};
 			before( function() {
 				this.clock = sinon.useFakeTimers();
@@ -230,7 +234,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 				this.clock.restore();
 			} );
 			it( "should cause child FSM to handle a _reset input", function() {
-				events[ 18 ].should.eql( {
+				events[ 22 ].should.eql( {
 					name: "transition",
 					data: {
 						fromState: "yellow",
@@ -242,7 +246,7 @@ describe( "Hierarchical machina.BehavioralFsm", function() {
 				} );
 			} );
 			it( "should emit a 'vehicles - green' event", function() {
-				events[ 19 ].should.eql( { name: "vehicles", data: { status: "green", client: client } } );
+				events[ 23 ].should.eql( { name: "vehicles", data: { status: "green", client: client } } );
 			} );
 		} );
 	} );