+++ spec/hierarchical.Fsm.spec.js
@@ -1,8 +1,12 @@
+var _ = require( "lodash" );
+var sinon = require( "sinon" );
+var machina = require( "../lib/machina.js" );
+var hierarchical = require( "./helpers/hierarchicalFsm.js" )( machina );
+
 describe( "Hierarchical machina.Fsm", function() {
 	describe( "when creating a hierarchy", function() {
 		var crosswalk;
 		var events = [];
-		var resetHandled = false;
 		before( function() {
 			crosswalk = hierarchical.crosswalkFactory( {
 				eventListeners: {
@@ -16,6 +20,7 @@ describe( "Hierarchical machina.Fsm", function() {
 		} );
 		it( "should report correct starting state for parent FSM", function() {
 			crosswalk.state.should.equal( "vehiclesEnabled" );
+			crosswalk.compositeState().should.equal( "vehiclesEnabled.green" );
 		} );
 		it( "should emit a 'pedestrians - do not walk' event", function() {
 			events[ 1 ].should.eql( { name: "pedestrians", data: { status: "Do Not Walk" } } );
@@ -24,7 +29,7 @@ describe( "Hierarchical machina.Fsm", function() {
 			crosswalk.states.vehiclesEnabled._child.instance.state.should.equal( "green" );
 		} );
 		it( "should issue reset input to child FSM of active parent state", function() {
-			events[ 2 ].should.eql( {
+			events[ 3 ].should.eql( {
 				name: "handling",
 				data: {
 					inputType: "_reset",
@@ -35,7 +40,7 @@ describe( "Hierarchical machina.Fsm", function() {
 			} );
 		} );
 		it( "should emit a 'vehicles - green' event", function() {
-			events[ 5 ].should.eql( { name: "vehicles", data: { status: "green" } } );
+			events[ 6 ].should.eql( { name: "vehicles", data: { status: "green" } } );
 		} );
 		it( "should not be listening to any events from child FSM of inactive parent state", function() {
 			_.any( events, function( item ) {
@@ -47,7 +52,6 @@ describe( "Hierarchical machina.Fsm", function() {
 		describe( "and the input originates from timer in child", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			before( function() {
 				this.clock = sinon.useFakeTimers();
 				crosswalk = hierarchical.crosswalkFactory( {
@@ -84,6 +88,7 @@ describe( "Hierarchical machina.Fsm", function() {
 					}
 				} );
 				crosswalk.states.vehiclesEnabled._child.instance.state.should.equal( "green-interruptible" );
+				crosswalk.compositeState().should.equal( "vehiclesEnabled.green-interruptible" );
 			} );
 			it( "should not change parent FSM's state", function() {
 				crosswalk.state.should.equal( "vehiclesEnabled" );
@@ -92,7 +97,6 @@ describe( "Hierarchical machina.Fsm", function() {
 		describe( "and the input originates from parent FSM", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			before( function() {
 				this.clock = sinon.useFakeTimers();
 				crosswalk = hierarchical.crosswalkFactory( {
@@ -126,6 +130,7 @@ describe( "Hierarchical machina.Fsm", function() {
 					}
 				} );
 				crosswalk.states.vehiclesEnabled._child.instance.state.should.equal( "yellow" );
+				crosswalk.compositeState().should.equal( "vehiclesEnabled.yellow" );
 			} );
 			it( "should emit a 'vehicles - yellow' event", function() {
 				events[ 2 ].should.eql( { name: "vehicles", data: { status: "yellow" } } );
@@ -134,7 +139,6 @@ describe( "Hierarchical machina.Fsm", function() {
 		describe( "and input isn't handled in child, but parent instead", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			before( function() {
 				this.clock = sinon.useFakeTimers();
 				crosswalk = hierarchical.crosswalkFactory( {
@@ -173,12 +177,13 @@ describe( "Hierarchical machina.Fsm", function() {
 					}
 				} );
 				crosswalk.state.should.equal( "pedestriansEnabled" );
+				crosswalk.compositeState().should.equal( "pedestriansEnabled.walking" );
 			} );
 			it( "should emit a 'vehicles - red' event", function() {
 				events[ 2 ].should.eql( { name: "vehicles", data: { status: "red" } } );
 			} );
 			it( "should have child FSM of active parent state handle _reset input", function() {
-				events[ 3 ].should.eql( {
+				events[ 4 ].should.eql( {
 					name: "handling",
 					data: {
 						inputType: "_reset",
@@ -189,13 +194,12 @@ describe( "Hierarchical machina.Fsm", function() {
 				} );
 			} );
 			it( "should emit a 'pedestrians - walk' event", function() {
-				events[ 6 ].should.eql( { name: "pedestrians", data: { status: "Walk" } } );
+				events[ 7 ].should.eql( { name: "pedestrians", data: { status: "Walk" } } );
 			} );
 		} );
 		describe( "and parent FSM transitions into previously held state", function() {
 			var crosswalk;
 			var events = [];
-			var resetHandled = false;
 			before( function() {
 				this.clock = sinon.useFakeTimers();
 				crosswalk = hierarchical.crosswalkFactory( {
@@ -213,7 +217,7 @@ describe( "Hierarchical machina.Fsm", function() {
 				this.clock.restore();
 			} );
 			it( "should cause child FSM to handle a _reset input", function() {
-				events[ 17 ].should.eql( {
+				events[ 21 ].should.eql( {
 					name: "transition",
 					data: {
 						fromState: "yellow",
@@ -224,7 +228,7 @@ describe( "Hierarchical machina.Fsm", function() {
 				} );
 			} );
 			it( "should emit a 'vehicles - green' event", function() {
-				events[ 18 ].should.eql( { name: "vehicles", data: { status: "green" } } );
+				events[ 22 ].should.eql( { name: "vehicles", data: { status: "green" } } );
 			} );
 		} );
 	} );